fix: reduce api calls to drastically increase performance
get_notes_info(deck_name) function: This function fetches all the notes in the given deck at once. It first finds all note ids with the findNotes API and then fetches their information with the notesInfo API. This replaces the note_exists function.

existing_notes variable: This is a list of all existing notes in the deck. We fetch them at the beginning of the create_anki_notes function, so we don't have to call the API for every single note we're trying to add.

existing_backs set: This is a set of all existing 'Back' values in the deck. We use a set here because checking membership in a set is faster than checking membership in a list (O(1) vs O(n)). We then check if the 'Back' of the note we're trying to add is in this set.

These changes significantly reduce the number of API calls made, which was the main cause of the performance issue.